refactor: 알림 도메인 패키지 구조 정리
알림 기준은 알림 기준끼리, 이메일 알림 설정은 이메일 알림 설정끼리 패키지로 정리!
