+++ src/main/java/com/alzzaipo/hexagonal/notification/application/port/in/criterion/RegisterNotificationCriterionUseCase.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.hexagonal.notification.application.port.in;
+package com.alzzaipo.hexagonal.notification.application.port.in.criterion;
 
 import com.alzzaipo.hexagonal.notification.application.port.dto.RegisterNotificationCriterionCommand;
