+++ src/main/java/com/alzzaipo/hexagonal/notification/application/service/criterion/DeleteNotificationCriterionService.java
@@ -1,31 +1,31 @@
-package com.alzzaipo.hexagonal.notification.application.service;
+package com.alzzaipo.hexagonal.notification.application.service.criterion;
 
 import com.alzzaipo.hexagonal.common.Uid;
 import com.alzzaipo.hexagonal.member.application.port.out.member.FindMemberPort;
-import com.alzzaipo.hexagonal.notification.application.port.dto.UpdateNotificationCriterionCommand;
-import com.alzzaipo.hexagonal.notification.application.port.in.UpdateNotificationCriterionUseCase;
-import com.alzzaipo.hexagonal.notification.application.port.out.FindNotificationCriterionPort;
-import com.alzzaipo.hexagonal.notification.application.port.out.UpdateNotificationCriterionPort;
-import com.alzzaipo.hexagonal.notification.domain.NotificationCriterion;
+import com.alzzaipo.hexagonal.notification.application.port.dto.DeleteNotificationCriterionCommand;
+import com.alzzaipo.hexagonal.notification.application.port.in.criterion.DeleteNotificationCriterionUseCase;
+import com.alzzaipo.hexagonal.notification.application.port.out.criterion.DeleteNotificationCriterionPort;
+import com.alzzaipo.hexagonal.notification.application.port.out.criterion.FindNotificationCriterionPort;
+import com.alzzaipo.hexagonal.notification.domain.criterion.NotificationCriterion;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class UpdateNotificationCriterionService implements UpdateNotificationCriterionUseCase {
+public class DeleteNotificationCriterionService implements DeleteNotificationCriterionUseCase {
 
     private final FindMemberPort findMemberPort;
     private final FindNotificationCriterionPort findNotificationCriterionPort;
-    private final UpdateNotificationCriterionPort updateNotificationCriterionPort;
+    private final DeleteNotificationCriterionPort deleteNotificationCriterionPort;
 
     @Override
-    public void updateNotificationCriterion(UpdateNotificationCriterionCommand command) {
+    public void deleteNotificationCriterion(DeleteNotificationCriterionCommand command) {
         validateMemberAndOwnership(command);
 
-        updateNotificationCriterionPort.updateNotificationCriterion(command);
+        deleteNotificationCriterionPort.deleteNotificationCriterion(command.getNotificationCriterionUID());
     }
 
-    private void validateMemberAndOwnership(UpdateNotificationCriterionCommand command) {
+    private void validateMemberAndOwnership(DeleteNotificationCriterionCommand command) {
         Uid memberUID = command.getMemberUID();
         Uid notificationCriterionUID = command.getNotificationCriterionUID();
