+++ src/main/java/com/alzzaipo/hexagonal/notification/application/service/criterion/FindMemberNotificationCriteriaService.java
@@ -1,11 +1,11 @@
-package com.alzzaipo.hexagonal.notification.application.service;
+package com.alzzaipo.hexagonal.notification.application.service.criterion;
 
 import com.alzzaipo.hexagonal.common.Uid;
 import com.alzzaipo.hexagonal.member.application.port.out.member.FindMemberPort;
 import com.alzzaipo.hexagonal.notification.application.port.dto.NotificationCriterionView;
-import com.alzzaipo.hexagonal.notification.application.port.in.FindMemberNotificationCriteriaQuery;
-import com.alzzaipo.hexagonal.notification.application.port.out.FindMemberNotificationCriteriaPort;
-import com.alzzaipo.hexagonal.notification.domain.NotificationCriterion;
+import com.alzzaipo.hexagonal.notification.application.port.in.criterion.FindMemberNotificationCriteriaQuery;
+import com.alzzaipo.hexagonal.notification.application.port.out.criterion.FindMemberNotificationCriteriaPort;
+import com.alzzaipo.hexagonal.notification.domain.criterion.NotificationCriterion;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
